Added NavigationView and DailySleepHeaderView
